+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8J1YQ9N8YDK5VK3W07XRY5C"
+        "01K8JEAJ3TPZCHFW87PNYXVV51"
     )
     r"""The unique identifier of the data source"""