+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8J1YQ9J40DQ73B3AK2CPSTA"
+        "01K8JEAJ3PR58PGPMMVC3721VR"
     )
     r"""The unique identifier of the data source"""