+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8J1YQ9MEBEA68EH43TZY5YF"
+        "01K8JEAJ3RP00AHYZ7D892DWAA"
     )
     r"""The unique identifier of the data source"""