+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -121,5 +121,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-27T05:23:34.900Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-27T08:59:46.066Z")
     r"""The date and time the resource was last updated"""