+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K8J1YQ9QPT3JME7QJQ27V2Z3"
+        "01K8JEAJ3XV27NBRVKXAQZXC2C"
     )
     r"""The unique identifier of the data source"""